Added support for unquoted String Values
In this version, if you set the setting "JsonOptions.AllowUnquotedKeys" to true it will also allow unquoted value strings. Unquoted Strings (in this version) end once we find a space, a coma a "}" or a "{".
Notice this captures the numeric values (since they might be a string) but if IsNumeric() it converts them back to number using VBA.Val().
Future version should clarify the setting: (a) rename the setting to clarify it works for both key+values or (b) create a new setting to allow unquote key/values separately.